Fix construction order of base::Thread members
Previously, the thread member was declared earlier, and was thus
constructed before the mutex. This led to a subtle race condition in
which the mutex could occasionally be used before it was properly
initialized, causing a segfault. base::Thread now takes care to start
the thread last, in order to avoid such problems.